refactor(GODT-2224): Allow user to control how to download attachments
Expose mechanism for a user of the library to control how the
attachments should be downloaded.

`AttachmentAllocator` allows the user to specify where the byte
buffers for the attachments should originate from.

`Scheduler` allows the user to control how the attachments should be
downloaded. `SequentialScheduler` and `ParallelScheduler` are provided
as reference implementations.