test: Make server status test more robust
The test HTTP client is not configured with any timeouts.
If the read fails, depending on when it happens, it can hang indefinitely.
This was unpredictable with the way the test was written before;
the message fetching was highly parallel. By replacing the message fetch
with a simple GetUser call which fails predictably, the test is made
more robust.